requirements.txt
fastapi==0.78.0
pydantic>=1.9.1,<2.0.0
python-dotenv>=0.20.0,<0.21.0
uvicorn>=0.18.2,<0.19.0
aiofiles>=0.8.0,<1.0.0
Jinja2>=3.1.2,<3.2.0
SQLAlchemy>=1.4.39,<1.5.0
pyjwt>=2.4.0,<2.5.0
pymysql>=1.0.2,<1.1.0